A: oauth_url persistence A: POST method for login
OAuth URL was requested several times for testig for a single server
instance. This has been made redundant by using a persistent oauth url
variable.

Since, we need to send data (pin) to the server, I had to change the GET
request for /auth/twitter/login to a POST request!