A: pinbox, new twitter window, axios post request
Twitter URL opens in new tab (for tabbed browsers or accordig to user
preference) and asks for authentication and provides a Pin.

Pin input has been added and it correctly sends the pin inputed by the
user to server.

Axios POST request sends the Pin and sets the access token and key in
local storage.