A: loginUser function, form only if token invalid
loginUser function in App now correctly fetches user details and sets
the user state value with username and profile url.

login with twitter button is only visible if the tokens are not present
otherwise the user is logged in automatically! Earlier, user may click
the button to login in the brief window that it gets. Hence resulting in
dual OAuth requests even when we have access tokens.