Fix slow Login Bug
The bug was caused because the loginUser method launched a promise and
returned control to handleSubmit without finishing the promise. Thus
user was queried before his auth had been setup in the backend.

This commit makes major changes in the codebase to fix this bug and also
to modify the filtering and fetching of tweets.

1. The loginUser method handles all the login scenarios, via
localStorage or otherwise with pin. The login component just displays
the form.

2. The tweets are no longer filtered or stored in the session in server,
instead, the job for handling tweets, querying for the correct tweets is
passed down to the client side.

The tweet filtering and querying has not been completely set for now.
The get more tweets button doen't work as expected. The button will be
deprecated and a loader will be added to autofetch tweets.